move model_dashboard to templates folder
