+++ ui/litellm-dashboard/src/app/page.tsx
@@ -8,7 +8,7 @@ import { Team } from "@/components/key_team_helpers/key_list"
 import Navbar from "@/components/navbar"
 import { ThemeProvider } from "@/contexts/ThemeContext"
 import UserDashboard from "@/components/user_dashboard"
-import ModelDashboard from "@/components/model_dashboard"
+import ModelDashboard from "@/components/templates/model_dashboard"
 import ViewUserDashboard from "@/components/view_users"
 import Teams from "@/components/teams"
 import Organizations from "@/components/organizations"
